Removed the forced jolokia depedency. Due to the way that the fat jar for the CorDapp is constructed after the dependency is forced, there are     duplicated files in the JAR. Furthermore jolokia is a potential     security risk as it increases the attack surfaces, and therefore     should only be included by a project if they wish to use it.
Determine whether to use Quasar from the lib directory (as per older
versions of Cordformation), or to pick up a specific version as a
dependency based on whether or not a group for quasar has been
specified.

Log the dependencies that will be included in the final CorDapp JAR file
to make it easier to see what the result of these plugins are.