CORDA-1915 CorDapp signed by default caused Jar task to be always out-of-sync, the signJar task was overwriting the Jar task output, delete the signJar task and move signing as part of the Jar task so it's output is unchanged by other task (#127)
Addition of the signJar Task running after jar task by CORDA-1915 has broken up-to-date validity of the first one. CorDapp signed by default caused Jar task to be always out-of-sync, the signJar task was overwriting the Jar task output, delete the signJar task and move signing as part of the Jar task so it's output is unchanged by other task.