Fixed bug with Gradle zip expansion
Removed list of `org.gradle.api.internal.file.copy.FileCopyDetailsInternal` that
was previosuly used to process files twice (once to compute the jar manifest and once to actually create the resulting jar archives).
Unfortunately, `FileCopyDetailsInternal` instances cannot be stored and must instead be used as soon as possible
because they are not guaranteed to be valid at a later stage.
This problem became clear with the usage of Gradle's `zipTree` and `tarTree` APIs that
lazily create an epxanded version of an archive in a temporary directory but, if the `FileCopyDetailsInternal`
instance isn't used immediately the archive entry isn't even extracted and cannot be referenced at a later stage since
the source archive is then closed.