CORE-924 fixed RandomAccessFile leak and race condition on lock file creation
`RandomAccessFile` is a `Closeable` and needs to be closed, otherwise it will leak an open file descriptor that could prevent the file from being deleted on some OSes (namely Windows).
There is also a race condition where multiple Flask processes are started, the lock file file doesn't exist so one process try to create it but between the `File.exists` call and the `Files.createFile` call another process creates it causing Files.createFile to fail with `java.nio.file.FileAlreadyExistsException`. Now, considering we never write to the lock file, we don't actually need to use `RandomAccessFile`, we can happily get by with a simple `FileChannel` instance (which we need to remember to close of course) and solve both of the aforementioned issues.