CORE-17494: Add check to CPB creation to ensure that 2 CPKs don't share a name (#596)
We are adding the ability for users to set an identifiable name for their own CPKs. However, we do not want 2 CPKs in the same CPB to have the same name, so we need to add a check in the CPB plugin to ensure that names are not reused within it.