CORE-17493: Add cordappName field to contract and workflow fields of cpk2 (#595)
In a cordapp CPK we have a field that sets a name for a cordapp that persists through versions. However, this is assigned automatically based on other values, so we are adding the ability for users to set this name for their own cordapp manually to allow for smoother upgrades where those other values may need to change, as well as to help make CPKs more identifiable to developers.